Remove powOf2 method
Without rounding and error correction it does not deliver the correct
value.
Unfortunately the precision at the error correction at the pow method
does also deliver incorrect values. It would require more tests and
better understanding about the math behind the way how the error
correction is done to deliver use-able results.
For now we decided to not use that library.